requirements.txt
Django==2.1.3
django-crispy-forms==1.7.2
django-filter==2.0.0
django-widget-tweaks==1.4.3
Faker==0.9.2
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
text-unidecode==1.2
